Preserved tavg when provided by station
Previously the code would always calculate TAvg as (TMax+Tmin)/2 even when it was provided by the station. This calculation will still occur should TAvg not be provided.

Now, when data is corrected, TAvg observations will be removed if the corresponding TMax or TMin observations are also missing or have been removed as outliers.